load sbv2 only once on mount
