+++ src/pages/_app.tsx
@@ -1,5 +1,4 @@
 import { AppProps } from "next/app";
-import Head from "next/head";
 import { FC } from "react";
 
 import { Layout } from "../components";
