+++ src/utils/obligation.ts
@@ -1,0 +1,1 @@
+export const a = 5;
